requirements.txt
browser-use
python-dotenv
anthropic
chromadb
openai
selenium
